adds an existence check for twitter image copying

+++ bkmkorg/twitter_data.py
@@ -259,6 +259,9 @@ class TwitterData:
         for x,y in zip(old_paths, new_paths):
             if exists(y) and not self.check_hashes(x,y):
                 raise Exception("File already exists: {} {}".format(x, y))
+            if not exists(x):
+                logging.warning("File Doesn't Exist: {}".format(x))
+                continue
             copyfile(x, y)
 
     def move_videos(self, new_dir):
